feat: convert the plugin to a class-based plugin
